requirements.txt
dwave-ocean-sdk==6.7.0
matplotlib==3.7.1
numpy==1.23.5
pandas==1.5.3
pulp==2.7.0
